+++ test/other/embind_tsgen_ignore_1.d.ts
@@ -26,9 +26,9 @@ declare namespace RuntimeExports {
 interface WasmModule {
   _pthread_self(): number;
   _main(_0: number, _1: number): number;
+  __embind_initialize_bindings(): void;
   __emscripten_tls_init(): number;
   __emscripten_proxy_main(_0: number, _1: number): number;
-  __embind_initialize_bindings(): void;
   __emscripten_thread_init(_0: number, _1: number, _2: number, _3: number, _4: number, _5: number): void;
   __emscripten_thread_crashed(): void;
   __emscripten_thread_exit(_0: number): void;